+++ src/datatypes/mod.rs
@@ -1,6 +1,7 @@
 pub mod commit;
 pub mod game_version;
 pub mod geolocation;
+pub mod hotkey;
 pub mod installation;
 pub mod server;
 pub mod value_sorted_map;
