+++ src/states/app.rs
@@ -9,22 +9,24 @@ use anyhow::Result;
 use crate::app::AppAction;
 use crate::config::AppConfig;
 use crate::constants::USER_AGENT;
+use crate::datatypes::hotkey::HotKey;
 use crate::states::events::EventsState;
 use crate::states::help::HelpState;
-use crate::states::help::HotKey;
 #[cfg(feature = "geolocation")]
 use crate::states::LocationsState;
-use crate::states::{CommitState, ServersState, VersionsState};
+use crate::views::commits::Commits;
+use crate::views::servers::Servers;
+use crate::views::versions::Versions;
 
 pub type TaskResult = Result<()>;
 
 pub struct AppState {
     pub config: AppConfig,
-    pub commits: Arc<RwLock<CommitState>>,
-    pub versions: Arc<RwLock<VersionsState>>,
+    pub commits: Commits,
+    //pub versions: Arc<RwLock<VersionsState>>,
     #[cfg(feature = "geolocation")]
     pub locations: Arc<RwLock<LocationsState>>,
-    pub servers: Arc<RwLock<ServersState>>,
+    pub servers: Servers,
     pub events: Arc<RwLock<EventsState>>,
 
     pub help: Mutex<HelpState>,
@@ -38,13 +40,13 @@ impl AppState {
     pub async fn new(config: AppConfig, panic_bool: Arc<AtomicBool>) -> Arc<Self> {
         #[cfg(feature = "geolocation")]
         let locations = Arc::new(RwLock::new(LocationsState::new(&config).await));
-        let versions = Arc::new(RwLock::new(VersionsState::new(&config).await));
-        let servers = Arc::new(RwLock::new(ServersState::new(&config).await));
+        let versions = Versions::new();
+        let servers = Servers::new();
         let events = Arc::new(RwLock::new(EventsState::new(&config).await));
 
         let instance = Arc::new(Self {
-            commits: Arc::new(RwLock::new(CommitState::new().await)),
-            versions: versions.clone(),
+            commits: Commits::new(),
+            //versions: versions.clone(),
             #[cfg(feature = "geolocation")]
             locations: locations.clone(),
             servers: servers.clone(),
@@ -61,44 +63,43 @@ impl AppState {
         });
 
         events.write().await.run(instance.clone()).await;
-        servers.write().await.run(instance.clone()).await;
+        servers.run(instance.clone()).await;
         #[cfg(feature = "geolocation")]
         locations.write().await.run(instance.clone()).await;
-        versions.write().await.run(instance.clone()).await;
-
+        //versions.write().await.run(instance.clone()).await;
         instance
     }
 
     pub async fn on_action(&self, action: &AppAction, app: Arc<AppState>) {
         log::debug!("action: {:?}", &action);
 
-        let f = match action {
-            AppAction::ConnectToServer { version, address } => Some(tokio::spawn(
-                VersionsState::launch(Arc::clone(&app), version.clone(), Some(address.clone())),
-            )),
-            AppAction::InstallVersion(version) => Some(tokio::spawn(VersionsState::install(
-                Arc::clone(&app),
-                version.clone(),
-            ))),
-            AppAction::LaunchVersion(version) => Some(tokio::spawn(VersionsState::launch(
-                Arc::clone(&app),
-                version.clone(),
-                None,
-            ))),
-            AppAction::AbortVersionInstallation(version) => Some(tokio::spawn(
-                VersionsState::abort_installation(Arc::clone(&app), version.clone()),
-            )),
-            AppAction::UninstallVersion(version) => Some(tokio::spawn(VersionsState::uninstall(
-                Arc::clone(&app),
-                version.clone(),
-            ))),
+        // let f = match action {
+        //     AppAction::ConnectToServer { version, address } => Some(tokio::spawn(
+        //         VersionsState::launch(Arc::clone(&app), version.clone(), Some(address.clone())),
+        //     )),
+        //     AppAction::InstallVersion(version) => Some(tokio::spawn(VersionsState::install(
+        //         Arc::clone(&app),
+        //         version.clone(),
+        //     ))),
+        //     AppAction::LaunchVersion(version) => Some(tokio::spawn(VersionsState::launch(
+        //         Arc::clone(&app),
+        //         version.clone(),
+        //         None,
+        //     ))),
+        //     AppAction::AbortVersionInstallation(version) => Some(tokio::spawn(
+        //         VersionsState::abort_installation(Arc::clone(&app), version.clone()),
+        //     )),
+        //     AppAction::UninstallVersion(version) => Some(tokio::spawn(VersionsState::uninstall(
+        //         Arc::clone(&app),
+        //         version.clone(),
+        //     ))),
 
-            _ => None,
-        };
+        //     _ => None,
+        // };
 
-        if let Some(f) = f {
-            self.watch_task(f).await;
-        }
+        // if let Some(f) = f {
+        //     self.watch_task(f).await;
+        // }
     }
 
     pub fn display_help(&self, view_name: &str, keys: &[HotKey]) {
