+++ src/views/commits/draw.rs
@@ -6,56 +6,17 @@ use tui::{
     layout::{Alignment, Constraint, Direction, Layout, Rect},
     style::{Color, Style},
     text::Text,
-    widgets::{Block, Borders, List, ListItem, ListState, Paragraph, Wrap},
+    widgets::{Block, Borders, List, ListItem, Paragraph, Wrap},
     Frame,
 };
 
-use crate::app::AppAction;
-use crate::input::UserInput;
-use crate::states::help::HotKey;
-use crate::states::{AppState, CommitState, StatelessList};
-use crate::views::{Drawable, HotKeys, InputProcessor, Named};
+use crate::states::AppState;
+use crate::views::Draw;
 
-pub struct CommitView {
-    // TODO:
-    //   - on 1st launch: fetch N latest commits, save latest hash
-    //   - on 2nd launch: read latest hash and fetch newer commits
-    loaded: bool,
-
-    state: StatelessList<ListState>,
-}
-
-impl CommitView {
-    pub fn new() -> Self {
-        Self {
-            loaded: false,
-            state: StatelessList::new(ListState::default(), false),
-        }
-    }
-}
+use super::Commits;
 
 #[async_trait::async_trait]
-impl Named for CommitView {
-    fn name(&self) -> String {
-        "Recent Commit List".to_owned()
-    }
-}
-
-impl HotKeys for CommitView {
-    fn hotkeys(&self) -> Vec<HotKey> {
-        self.state.hotkeys()
-    }
-}
-
-#[async_trait::async_trait]
-impl InputProcessor for CommitView {
-    async fn on_input(&mut self, input: &UserInput, app: Arc<AppState>) -> Option<AppAction> {
-        self.state.on_input(input, app.commits.read().await.count())
-    }
-}
-
-#[async_trait::async_trait]
-impl Drawable for CommitView {
+impl Draw for Commits {
     async fn draw(
         &mut self,
         f: &mut Frame<CrosstermBackend<io::Stdout>>,
@@ -67,7 +28,7 @@ impl Drawable for CommitView {
             .constraints([Constraint::Length(60), Constraint::Min(0)])
             .split(area);
 
-        let commits = &app.commits.read().await.items;
+        let commits = &self.state.read().await.items;
 
         let items: Vec<ListItem> = commits
             .iter()
@@ -83,9 +44,9 @@ impl Drawable for CommitView {
             )
             .highlight_style(Style::default().bg(Color::DarkGray));
 
-        f.render_stateful_widget(list, chunks[0], &mut self.state.state);
+        f.render_stateful_widget(list, chunks[0], &mut self.selection.state);
 
-        if let Some(i) = self.state.state.selected() {
+        if let Some(i) = self.selection.state.selected() {
             let chunks = Layout::default()
                 .direction(Direction::Vertical)
                 .constraints([Constraint::Length(3), Constraint::Min(0)])
@@ -120,7 +81,7 @@ impl Drawable for CommitView {
 
         if !self.loaded {
             if !app.config.offline {
-                app.watch_task(tokio::spawn(CommitState::load(Arc::clone(&app))))
+                app.watch_task(tokio::spawn(self.load(Arc::clone(&app))))
                     .await;
             }
