+++ src/views/commits/state.rs
@@ -1,0 +1,11 @@
+use crate::datatypes::commit::Commit;
+
+pub struct State {
+    pub items: Vec<Commit>,
+}
+
+impl State {
+    pub fn new() -> Self {
+        Self { items: Vec::new() }
+    }
+}
