+++ src/views/servers/state.rs
@@ -1,0 +1,9 @@
+pub struct State {
+    pub items: Vec<Server>,
+}
+
+impl State {
+    pub fn new() -> Self {
+        Self { items: Vec::new() }
+    }
+}
